+++ src/App.jsx
@@ -1,10 +1,11 @@
 import React from "react"
+import SponsorshipSection from './components/SponsorshipSection';
 
 function App() {
 
   return (
     <>
-      
+      <SponsorshipSection />;
     </>
   )
 }
